+++ src/language/diagnose.ts
@@ -27,7 +27,6 @@ export class landiagnose{
         let flag =2
         this.asmerror=0
         this.asmwarn=0
-        //console.log(text,info)
         let MASMorTASM:string|undefined
         if (ASM) MASMorTASM=ASM
         else{
@@ -164,7 +163,7 @@ private tasmdiagpush(severity:number,line:number,msg:string,text:string,related?
                 {
                     let related=this.TasmMacroRelated(fileuri,text,macroname,macroline,msg)
                     let line=parseInt(line_get)
-                    this.tasmdiagpush(severity,line,msg,text,related)
+                    this.tasmdiagpush(severity,line,msg+" (in macro \""+macroname+"\" "+line_get+")",text,related)
                     delete allmsg[i]
                 }
             }